src: stop using size_t everywhere. use ints where possible.
See: https://www.youtube.com/watch?v=Puio5dly9N8&t=2558s

+++ include/torsion/rsa.h
@@ -99,11 +99,11 @@ extern "C" {
 TORSION_EXTERN int
 rsa_privkey_generate(unsigned char *out,
                      size_t *out_len,
-                     unsigned long bits,
+                     unsigned int bits,
                      uint64_t exp,
                      const unsigned char *entropy);
 
-TORSION_EXTERN size_t
+TORSION_EXTERN unsigned int
 rsa_privkey_bits(const unsigned char *key, size_t key_len);
 
 TORSION_EXTERN int
@@ -128,7 +128,7 @@ rsa_pubkey_create(unsigned char *out,
                   const unsigned char *key,
                   size_t key_len);
 
-TORSION_EXTERN size_t
+TORSION_EXTERN unsigned int
 rsa_pubkey_bits(const unsigned char *key, size_t key_len);
 
 TORSION_EXTERN int
@@ -233,7 +233,7 @@ rsa_veil(unsigned char *out,
          size_t *out_len,
          const unsigned char *msg,
          size_t msg_len,
-         size_t bits,
+         unsigned int bits,
          const unsigned char *key,
          size_t key_len,
          const unsigned char *entropy);
@@ -243,7 +243,7 @@ rsa_unveil(unsigned char *out,
            size_t *out_len,
            const unsigned char *msg,
            size_t msg_len,
-           size_t bits,
+           unsigned int bits,
            const unsigned char *key,
            size_t key_len);
